==> 01_apify.ipynb <==
pip install apify_client

txt = pd.read_csv('place_ids.txt')

## Getting reviews for muliple place ids

import os
import pandas as pd
from apify_client import ApifyClient

# ==== SET INPUT PARAMETERS HERE ====

for i in txt['place_id'][139:].values:
    place_id =i   # 👈 replace with your place_id
    #APIFY_API_TOKEN = os.getenv("APIFY_API_TOKEN", "your_apify_token_here")  # replace if not set in env
    output_dir = "./data"
     #"account1": "apify_api_i2hSYSzqEa9tw2rqFG0jf80veMbHyX14OdID",
     #"account2": "apify_api_XQdxmw8RxbgoI8cZ1v84SxRoDVKH200zrQBs",
     #"account3": "apify_api_wddOUXwe0FIEKOMmxxhqpMrOmwBZYw2GyTIa"
    # ===================================
    
    # Initialize Apify client
    client = ApifyClient("apify_api_wddOUXwe0FIEKOMmxxhqpMrOmwBZYw2GyTIa")
    
    
    run_input = {
        "startUrls": [{ "url": f"https://www.google.com/maps/place/?q=place_id:{place_id}&hl=en" }],
        "maxReviews": 500,
        "reviewsSort": "newest",
        "language": "en",
        "reviewsOrigin": "all",
        "personalData": True,
               # Force UI + details in English
        "reviewsLanguage": "en",   
    }
        
    
    print(f"Starting Apify run for place_id={place_id}...")
    
    # Run the scraper actor
    
    run = client.actor("Xb8osYTtOjlsgI6k9").call(run_input=run_input)
    
    # Fetch dataset items
    dataset_items = list(client.dataset(run["defaultDatasetId"]).iterate_items())
    print(f"Retrieved {len(dataset_items)} items from Apify")
    
    # Convert results into Pandas DataFrame
    if dataset_items:
        df = pd.DataFrame(dataset_items)
        
        # Save CSV
        os.makedirs(output_dir, exist_ok=True)
        csv_file = os.path.join(output_dir, f"reviews_{place_id}.csv")
        df.to_csv(csv_file, index=False, encoding="utf-8-sig")
        
        print(f"Saved {len(df)} reviews to {csv_file}")
    else:
        print("No reviews found")



### Review for Single PLace id

import os
import pandas as pd
from apify_client import ApifyClient

# ==== SET INPUT PARAMETERS HERE ====


place_id ='ChIJHw31tJyFi0cRnUZQLOgt5Bk'   # 👈 replace with your place_id
#APIFY_API_TOKEN = os.getenv("APIFY_API_TOKEN", "your_apify_token_here")  # replace if not set in env

 #"account1": "apify_api_i2hSYSzqEa9tw2rqFG0jf80veMbHyX14OdID",
 #"account2": "apify_api_XQdxmw8RxbgoI8cZ1v84SxRoDVKH200zrQBs",
 #"account3": "apify_api_wddOUXwe0FIEKOMmxxhqpMrOmwBZYw2GyTIa"
# ===================================

# Initialize Apify client
client = ApifyClient("apify_api_wddOUXwe0FIEKOMmxxhqpMrOmwBZYw2GyTIa")


run_input = {
    "startUrls": [{ "url": f"https://www.google.com/maps/place/?q=place_id:{place_id}&hl=en" }],
    "maxReviews": 500,
    "reviewsSort": "newest",
    "language": "en",
    "reviewsOrigin": "all",
    "personalData": True,
           # Force UI + details in English
    "reviewsLanguage": "en",   
}



print(f"Starting Apify run for place_id={place_id}...")

# Run the scraper actor

run = client.actor("Xb8osYTtOjlsgI6k9").call(run_input=run_input)

# Fetch dataset items
dataset_items = list(client.dataset(run["defaultDatasetId"]).iterate_items())
print(f"Retrieved {len(dataset_items)} items from Apify")

# Convert results into Pandas DataFrame
if dataset_items:
    df = pd.DataFrame(dataset_items)
    
    # Save CSV
    os.makedirs(output_dir, exist_ok=True)
    csv_file = os.path.join( f"reviews_{place_id}.csv")
    df.to_csv(csv_file, index=False, encoding="utf-8-sig")
    
    print(f"Saved {len(df)} reviews to {csv_file}")
else:
    print("No reviews found")



df.to_csv('competitor_review.csv', index=False, encoding="utf-8-sig")

df
